VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@3465 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/protocol.cpp
@@ -168,7 +168,7 @@ const char *genpwdhash(const char *name, const char *pwd, int salt)
     static string temp;
     s_sprintf(temp)("%s %d %s %s %d", pwd, salt, name, pwd, PROTOCOL_VERSION);
     tiger::hashval hash;
-    tiger::hash((uchar *)temp, strlen(temp), hash);
+    tiger::hash((uchar *)temp, (int)strlen(temp), hash);
     s_sprintf(temp)("%llx %llx %llx", hash.chunks[0], hash.chunks[1], hash.chunks[2]);
     return temp;
 }
